dice-protocols: avid: move structure for hardware specification
This commit moves implementation of hardware specification for Avid Mbox
3 Pro from runtime crate to protocol crate.
